refactor: move event type to enums
